[docs] Update FC4 diagrams of fc4-tool
They were out of date since we recently removed the Node renderer that
executed as a discrete OS process.

As long as I was updating the diagrams themselves I took a moment to fix
the aspect ratio of the container diagram as rendered on the home page
of the FC4 website; it’s been off for awhile.