Update CCI config: change param os string -> enum
CCI = CircleCI

In d5ec078b00429 I changed this param from an enum to a string, so that
the values could be set in the executor (as env vars) and thereby passed
to the install-sys-deps command as a parameter. That worked, but it lost
the rigor and error-checking that an enum provides. So I thought I’d see
if I could restore it.

Just changing the type back to an enum wouldn’t work; the command
`circleci config validate` would fail, saying that "$OS" was not a valid
enum value. Which makes sense. So in order to make this command param
an enum, I had to also add it to the jobs that use the command, and then
set the values for the job param in the workflow where the jobs are
"called".

This is a little repetitive and makes the workflow a little harder to
read, but I suspect it’s worth doing for the additional rigor this
provides.