START: Simplify repo structure (#226)
As per #226, we’re redefining FC4 as a tool rather than a framework.

Until now, FC4 has been conceived of as a framework comprising a tool
and a methodology. The repo therefore contained the tool in the dir
`tool` and the methodology in `docs/methodology`.

Going forward, it’s simply a tool; the methodology will become the user
manual for the tool.

Therefore we can and should simplify the structure of the repo/project;
we should move the contents of `tool` out into the root (or wherever is
appropriate) and then delete `tool`.

This is the first step of that work.

Refs #226

WIP++

WIP++